Removed backend call
