+++ engine/src/test/scala/engine/backend/jes/JesBackendSpec.scala
@@ -149,9 +149,7 @@ class JesBackendSpec extends FlatSpec with Matchers with Mockito with BeforeAndA
     // This should only ever be used in this test to grab some locallyQualifiedInputs. So leave the rest null:
     val jobDescriptor = BackendCallJobDescriptor(null, null, inputs)
 
-    val mockedBackendCall = mock[JesJobDescriptor]
-    mockedBackendCall.jobDescriptor returns jobDescriptor
-    val mappedInputs: CallInputs = new JesBackend(actorSystem).adjustInputPaths(mockedBackendCall.jobDescriptor)
+    val mappedInputs: CallInputs = new JesBackend(actorSystem).adjustInputPaths(jobDescriptor)
 
     mappedInputs.get(stringKey).get match {
       case WdlString(v) => assert(v.equalsIgnoreCase(stringVal.value))
@@ -378,10 +376,10 @@ class JesBackendSpec extends FlatSpec with Matchers with Mockito with BeforeAndA
       workflowOptions = """ {"jes_gcs_root": "gs://path/to/gcs_root"} """
     )).copy(wfContext = new WorkflowContext("gs://path/to/gcs_root")).copy(
       fileSystems = List(GcsFileSystem.defaultGcsFileSystem, FileSystems.getDefault)
-    )
+    ).copy(backend = jesBackend)
 
     val call = wd.namespace.workflow.findCallByName("hello").get
-    val backendCall = jesBackend.bindCall(BackendCallJobDescriptor(wd, BackendCallKey(call, None, 1)))
+    val backendCall = BackendCallJobDescriptor(wd, BackendCallKey(call, None, 1))
     val stdoutstderr = backendCall.stdoutStderr
 
     stdoutstderr.stdout shouldBe WdlFile("gs://path/to/gcs_root/hello/e6236763-c518-41d0-9688-432549a8bf7c/call-hello/hello-stdout.log")
@@ -399,9 +397,9 @@ class JesBackendSpec extends FlatSpec with Matchers with Mockito with BeforeAndA
       workflowOptions = """ {"jes_gcs_root": "gs://path/to/gcs_root"} """
     )).copy(wfContext = new WorkflowContext("gs://path/to/gcs_root")).copy(
       fileSystems = List(GcsFileSystem.defaultGcsFileSystem, FileSystems.getDefault)
-    )
+    ).copy(backend = jesBackend)
     val call = wd.namespace.workflow.findCallByName("B").get
-    val backendCall = jesBackend.bindCall(BackendCallJobDescriptor(wd, BackendCallKey(call, Some(2), 1)))
+    val backendCall = BackendCallJobDescriptor(wd, BackendCallKey(call, Some(2), 1))
     val stdoutstderr = backendCall.stdoutStderr
 
     stdoutstderr.stdout shouldBe WdlFile("gs://path/to/gcs_root/w/e6236763-c518-41d0-9688-432549a8bf7c/call-B/shard-2/B-2-stdout.log")
